build.sh: enforce 2nd CEF copy to our desired revision
The CEF build machinery creates a second checkout of CEF under the
chrome source tree. The natural way for keeping them both in sync
would be just using a symlink, but that's confusing other parts of
the build machinery (eg. will detect the wrong source root dir, etc),
so it really needs to be an actual copy.

Calling git to check out the exact revision right before build, so
we don't need to do it every single time while working on the CEF
code base.

Downside: the changes in the dev repo first need to be pushed, before
calling the build.sh script.

TDB: find a better solution (maybe rsync ?)

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>